Improved score system to give points for each enemy damage, not only when you kill it. Adjusted points according with original 1942.

+++ Assets/scripts/Fire.cs
@@ -10,8 +10,8 @@ public class Fire : MonoBehaviour {
     private float speed = 3.0f;
 
     [SerializeField]
-    private float power;
-    public float Power
+    private int power;
+    public int Power
     {
         get { return this.power; }
     }
